refactor analyse-password.js 🥝 ..
